gpt_academic/requirements.txt
https://public.agent-matrix.com/publish/gradio-3.32.14-py3-none-any.whl
fastapi==0.110
gradio-client==0.8
pypdf2==2.12.1
httpx<=0.25.2
zhipuai==2.0.1
tiktoken>=0.3.3
requests[socks]
pydantic==2.9.2
protobuf==3.20
transformers>=4.27.1,<4.42
scipdf_parser>=0.52
spacy==3.7.4
anthropic>=0.18.1
python-markdown-math
pymdown-extensions>=10.14
websocket-client
beautifulsoup4
prompt_toolkit
latex2mathml
python-docx
mdtex2html
dashscope
pyautogen
colorama
Markdown
pygments
edge-tts>=7.0.0
pymupdf
openai
rjsmin
PyMySQL>=1.0.2
loguru
loguru
arxiv
numpy
rich


llama-index-core==0.10.68
llama-index-legacy==0.9.48
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-embeddings-azure-openai==0.1.10
llama-index-embeddings-openai==0.1.10
llama-parse==0.4.9
mdit-py-plugins>=0.3.3
linkify-it-py==2.0.3